change dotnet ver

+++ Core/CleanArchitecture.Domain/ViewModels/ProductView.cs
@@ -3,6 +3,5 @@
 public class ProductView
 {
     public string Name { get; set; } = null!;
-
-    public decimal Price { get; set; }
+    public float Price { get; set; }
 }
